requirements.txt
requests
python-dotenv
flask
flask-sqlalchemy
flask-login
gunicorn
psycopg2-binary
Flask-SQLAlchemy==2.1
flask_marshmallow
marshmallow_sqlalchemy